requirements.txt
# ── Core Framework ───────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database ─────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.5
alembic==1.13.3

# ── Background Tasks ────────────────────────────────────
celery[redis]==5.4.0
redis==5.1.1

# ── AI / Embeddings ─────────────────────────────────────
sentence-transformers==3.1.1
torch==2.4.1
numpy==1.26.4

# ── HTTP Client (for Ollama calls) ──────────────────────
httpx==0.27.2

# ── Git Operations ──────────────────────────────────────
GitPython==3.1.43

# ── Testing ──────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2

# ── Dev / Linting ────────────────────────────────────────
python-dotenv==1.0.1
